Checksum reconciled Instances and Bindings (#582)
* Add checksum field to InstanceSpec

* Add checksum to Binding

* Regenerate code

Regenerate code

more generation

+++ test/integration/clientset_test.go
@@ -533,6 +533,9 @@ func testInstanceClient(client servicecatalogclient.Interface, name string) erro
 	if e, a := readyConditionTrue, instanceServer.Status.Conditions[0]; !reflect.DeepEqual(e, a) {
 		return fmt.Errorf("Didn't get matching ready conditions:\nexpected: %v\n\ngot: %v", e, a)
 	}
+	if instanceServer.Spec.Checksum == nil {
+		return fmt.Errorf("Checksum should have been set after updating ready condition to true")
+	}
 
 	err = instanceClient.Delete(name, &v1.DeleteOptions{})
 	if nil != err {
@@ -691,6 +694,9 @@ func testBindingClient(client servicecatalogclient.Interface, name string) error
 	}
 	if e, a := readyConditionTrue, bindingServer.Status.Conditions[0]; !reflect.DeepEqual(e, a) {
 		return fmt.Errorf("Didn't get matching ready conditions:\nexpected: %v\n\ngot: %v", e, a)
+	}
+	if bindingServer.Spec.Checksum == nil {
+		return fmt.Errorf("Checksum should have been set after updating ready condition to true")
 	}
 
 	if err = bindingClient.Delete(name, &v1.DeleteOptions{}); nil != err {
